Reuse shared workflow run console
