+++ apps/aevatar-console-web/src/pages/team-member-workflow-studio/components/WorkflowStudioDraftRunPanel.tsx
@@ -1,14 +1,5 @@
-import {
-  CloseOutlined,
-  FileImageOutlined,
-  FileOutlined,
-  PaperClipOutlined,
-  PlayCircleOutlined,
-} from "@ant-design/icons";
-import { Button, Input, Tooltip, Typography } from "antd";
-import React from "react";
-import { t } from "@/shared/i18n/messages";
-import WorkflowStudioSidePanel from "./WorkflowStudioSidePanel";
+import React from 'react';
+import WorkflowRunInputPanel from '@/shared/workflows/WorkflowRunInputPanel';
 
 type WorkflowStudioDraftRunPanelProps = {
   readonly acceptedFileTypes?: string;
@@ -26,380 +17,29 @@ type WorkflowStudioDraftRunPanelProps = {
   readonly width?: number;
 };
 
-const DEFAULT_ACCEPTED_FILE_TYPES = [
-  "image/png",
-  "image/jpeg",
-  "image/webp",
-  "audio/mpeg",
-  "audio/wav",
-  "audio/wave",
-  "audio/x-wav",
-  "video/mp4",
-  "application/pdf",
-  "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
-  "text/csv",
-  "text/plain",
-  "text/markdown",
-  "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
-].join(",");
-
-const attachmentToolbarStyle: React.CSSProperties = {
-  alignItems: "center",
-  display: "flex",
-  flexWrap: "wrap",
-  gap: 8,
-  minWidth: 0,
-};
-
-const attachmentListStyle: React.CSSProperties = {
-  alignItems: "center",
-  display: "flex",
-  flex: "1 1 220px",
-  flexWrap: "wrap",
-  gap: 6,
-  minWidth: 0,
-};
-
-const attachmentChipStyle: React.CSSProperties = {
-  alignItems: "center",
-  background: "#f8fafc",
-  border: "1px solid #dbe3ee",
-  borderRadius: 8,
-  color: "#334155",
-  display: "inline-flex",
-  gap: 6,
-  maxWidth: "100%",
-  minHeight: 28,
-  minWidth: 0,
-  padding: "3px 5px 3px 8px",
-};
-
-const attachmentNameStyle: React.CSSProperties = {
-  fontSize: 12,
-  fontWeight: 700,
-  lineHeight: "16px",
-  maxWidth: 168,
-  overflow: "hidden",
-  textOverflow: "ellipsis",
-  whiteSpace: "nowrap",
-};
-
-const attachmentMetaStyle: React.CSSProperties = {
-  color: "#64748b",
-  flex: "0 0 auto",
-  fontSize: 11,
-  lineHeight: "14px",
-};
-
-const attachmentEmptyStyle: React.CSSProperties = {
-  color: "#64748b",
-  fontSize: 12,
-  lineHeight: "18px",
-};
-
-const fileDropZoneStyle: React.CSSProperties = {
-  alignItems: "center",
-  background: "#fbfcfe",
-  border: "1px dashed #cbd5e1",
-  borderRadius: 8,
-  color: "#475569",
-  display: "grid",
-  gap: 10,
-  justifyItems: "start",
-  minHeight: 84,
-  padding: "12px",
-};
-
-function formatFileSize(bytes: number): string {
-  if (!Number.isFinite(bytes) || bytes <= 0) {
-    return "0 B";
-  }
-
-  if (bytes < 1024) {
-    return `${bytes} B`;
-  }
-
-  const kib = bytes / 1024;
-  if (kib < 1024) {
-    return `${kib.toFixed(kib >= 10 ? 0 : 1)} KB`;
-  }
-
-  const mib = kib / 1024;
-  return `${mib.toFixed(mib >= 10 ? 0 : 1)} MB`;
-}
-
-function getFileIcon(file: File): React.ReactNode {
-  return file.type.startsWith("image/") ? <FileImageOutlined /> : <FileOutlined />;
-}
-
-function getAttachmentKey(file: File): string {
-  return [
-    file.name,
-    file.size,
-    file.lastModified,
-    file.type || "file",
-  ].join(":");
-}
-
-const WorkflowStudioDraftRunPanel: React.FC<WorkflowStudioDraftRunPanelProps> = ({
-  acceptedFileTypes = DEFAULT_ACCEPTED_FILE_TYPES,
+const WorkflowStudioDraftRunPanel: React.FC<
+  WorkflowStudioDraftRunPanelProps
+> = ({
+  acceptedFileTypes,
   canRun,
   disabledReason,
   files = [],
   onFilesAdd,
   onFileRemove,
-  onClose,
-  onRun,
-  onRunMessageChange,
-  open,
-  pending,
-  runMessage,
-  width = 420,
-}) => {
-  const fileInputRef = React.useRef<HTMLInputElement | null>(null);
-  const inputLocked = pending;
-
-  if (!open) {
-    return null;
-  }
-
-  const addFiles = (nextFiles: readonly File[]) => {
-    if (nextFiles.length > 0) {
-      onFilesAdd?.(nextFiles);
-    }
-  };
-
-  return (
-    <WorkflowStudioSidePanel
-      ariaLabel={t(
-        "teamMemberWorkflowStudio.draftRunPanel.sectionAria",
-        "Draft run panel",
-      )}
-      bodyStyle={{
-        display: "flex",
-        flexDirection: "column",
-        gap: 0,
-        overflow: "hidden",
-        padding: 0,
-      }}
-      closeAriaLabel={t(
-        "teamMemberWorkflowStudio.draftRunPanel.closeAria",
-        "Close draft run panel",
-      )}
-      onClose={onClose}
-      title={
-        <span style={{ alignItems: "center", display: "inline-flex", gap: 8 }}>
-          <PlayCircleOutlined />
-          <span>{t("teamMemberWorkflowStudio.draftRunPanel.title", "Draft run")}</span>
-        </span>
-      }
-      width={width}
-    >
-      <div
-        style={{
-          alignContent: "start",
-          display: "grid",
-          flex: "1 1 auto",
-          gap: 28,
-          minHeight: 0,
-          overflow: "auto",
-          padding: "28px 24px 24px",
-        }}
-      >
-        <section
-          style={{
-            display: "grid",
-            gap: 12,
-          }}
-        >
-          <div style={{ display: "grid", gap: 4 }}>
-            <Typography.Text strong>
-              {t(
-                "teamMemberWorkflowStudio.draftRunPanel.messageLabel",
-                "Draft run input",
-              )}
-            </Typography.Text>
-            <Typography.Text style={{ color: "#64748b" }}>
-              {t(
-                "teamMemberWorkflowStudio.draftRunPanel.emptyInputHint",
-                "Leave blank to run this draft without user input.",
-              )}
-            </Typography.Text>
-          </div>
-          <Input.TextArea
-            aria-label={t(
-              "teamMemberWorkflowStudio.draftRunPanel.messageLabel",
-              "Draft run input",
-            )}
-            autoSize={{ minRows: 7, maxRows: 10 }}
-            disabled={inputLocked}
-            onChange={(event) => onRunMessageChange(event.target.value)}
-            placeholder={t(
-              "teamMemberWorkflowStudio.draftRunPanel.messagePlaceholder",
-              "Optional input sent to this workflow draft run",
-            )}
-            style={{ fontSize: 15 }}
-            value={runMessage}
-          />
-        </section>
-
-        <section
-          style={{
-            display: "grid",
-            gap: 10,
-          }}
-        >
-          <div style={{ display: "grid", gap: 4 }}>
-            <Typography.Text strong>
-              {t("teamMemberWorkflowStudio.draftRunPanel.filesLabel", "Files")}
-            </Typography.Text>
-            <Typography.Text style={{ color: "#64748b" }}>
-              {t(
-                "teamMemberWorkflowStudio.draftRunPanel.filesHint",
-                "Attach files for this draft run.",
-              )}
-            </Typography.Text>
-          </div>
-          <input
-            ref={fileInputRef}
-            aria-label={t(
-              "teamMemberWorkflowStudio.draftRunPanel.attachFilesInput",
-              "Attach files",
-            )}
-            accept={acceptedFileTypes}
-            data-testid="draft-run-file-input"
-            multiple
-            onChange={(event) => {
-              addFiles(Array.from(event.target.files ?? []));
-              event.currentTarget.value = "";
-            }}
-            style={{ display: "none" }}
-            type="file"
-          />
-          <div
-            data-testid="draft-run-file-drop-zone"
-            onDragOver={(event) => {
-              event.preventDefault();
-              event.dataTransfer.dropEffect = inputLocked ? "none" : "copy";
-            }}
-            onDrop={(event) => {
-              event.preventDefault();
-              if (!inputLocked) {
-                addFiles(Array.from(event.dataTransfer.files ?? []));
-              }
-            }}
-            style={fileDropZoneStyle}
-          >
-            <div style={attachmentToolbarStyle}>
-              <Tooltip
-                title={t(
-                  "teamMemberWorkflowStudio.draftRunPanel.attachFiles",
-                  "Attach files",
-                )}
-              >
-                <Button
-                  aria-label={t(
-                    "teamMemberWorkflowStudio.draftRunPanel.attachFilesButton",
-                    "Add files",
-                  )}
-                  disabled={inputLocked}
-                  icon={<PaperClipOutlined />}
-                  onClick={() => fileInputRef.current?.click()}
-                >
-                  {t(
-                    "teamMemberWorkflowStudio.draftRunPanel.addFiles",
-                    "Add files",
-                  )}
-                </Button>
-              </Tooltip>
-              <Typography.Text style={{ color: "#64748b", fontSize: 12 }}>
-                {t(
-                  "teamMemberWorkflowStudio.draftRunPanel.dropFiles",
-                  "Drop files here",
-                )}
-              </Typography.Text>
-            </div>
-            <div style={attachmentListStyle}>
-              {files.length === 0 ? (
-                <span style={attachmentEmptyStyle}>
-                  {t(
-                    "teamMemberWorkflowStudio.draftRunPanel.noFilesAttached",
-                    "No files attached",
-                  )}
-                </span>
-              ) : (
-                files.map((file, index) => (
-                  <span
-                    key={getAttachmentKey(file)}
-                    data-testid="draft-run-file-chip"
-                    style={attachmentChipStyle}
-                  >
-                    {getFileIcon(file)}
-                    <Tooltip title={`${file.name} · ${file.type || "file"} · ${formatFileSize(file.size)}`}>
-                      <span style={attachmentNameStyle}>{file.name}</span>
-                    </Tooltip>
-                    <span style={attachmentMetaStyle}>{formatFileSize(file.size)}</span>
-                    <Button
-                      aria-label={t(
-                        "teamMemberWorkflowStudio.draftRunPanel.removeFile",
-                        "Remove {name}",
-                        { name: file.name },
-                      )}
-                      disabled={inputLocked}
-                      icon={<CloseOutlined />}
-                      onClick={() => onFileRemove?.(index)}
-                      size="small"
-                      type="text"
-                    />
-                  </span>
-                ))
-              )}
-            </div>
-          </div>
-          <Typography.Text style={{ color: "#64748b", fontSize: 12 }}>
-            {t(
-              "teamMemberWorkflowStudio.draftRunPanel.filesLimitHint",
-              "Images, documents, audio, video, CSV, and text files up to 10 MB.",
-            )}
-          </Typography.Text>
-        </section>
-      </div>
-
-      <div
-        style={{
-          borderTop: "1px solid #e5e7eb",
-          display: "grid",
-          gap: 10,
-          padding: "20px 24px 24px",
-        }}
-      >
-        <Button
-          disabled={!canRun}
-          icon={<PlayCircleOutlined />}
-          loading={pending}
-          onClick={onRun}
-          size="large"
-          style={{
-            boxShadow: canRun ? "0 12px 24px rgba(15, 23, 42, 0.14)" : undefined,
-            height: 54,
-            width: "100%",
-          }}
-          title={canRun ? undefined : disabledReason}
-          type="primary"
-        >
-          {t(
-            "teamMemberWorkflowStudio.draftRunPanel.startDraftRun",
-            "Start draft run",
-          )}
-        </Button>
-        {!canRun && disabledReason ? (
-          <Typography.Text style={{ color: "#6b7280", fontSize: 12 }}>
-            {disabledReason}
-          </Typography.Text>
-        ) : null}
-      </div>
-    </WorkflowStudioSidePanel>
-  );
-};
+  ...props
+}) => (
+  <WorkflowRunInputPanel
+    {...props}
+    canRun={canRun}
+    disabledReason={disabledReason}
+    variant={{
+      acceptedFileTypes,
+      files,
+      kind: 'draft',
+      onFilesAdd: (nextFiles) => onFilesAdd?.(nextFiles),
+      onFileRemove: (index) => onFileRemove?.(index),
+    }}
+  />
+);
 
 export default WorkflowStudioDraftRunPanel;
